+++ rest_framework_api_key/models.py
@@ -36,7 +36,7 @@ class APIKeyManager(models.Manager):
         return True
 
 
-class APIKey(models.Model):
+class BaseAPIKey(models.Model):
     objects = APIKeyManager()
 
     id = models.CharField(max_length=100, unique=True, primary_key=True)
@@ -67,6 +67,7 @@ class APIKey(models.Model):
     )
 
     class Meta:  # noqa
+        abstract = True
         ordering = ("-created",)
         verbose_name = "API key"
         verbose_name_plural = "API keys"
@@ -105,3 +106,7 @@ class APIKey(models.Model):
 
     def __str__(self) -> str:
         return str(self.name)
+
+
+class APIKey(BaseAPIKey):
+    pass